+++ src/components/Root/CommandPickerDetailed.tsx
