+++ src/components/Root/CommandPickerPanel.tsx
