+++ src/components/Root/ResultDialog.tsx
